code refactory for better maintanability
