+++ astonish/core/format_handler.py
@@ -40,8 +40,8 @@ async def execute_tool(
                 if not tool_schema_def:
                     raise ValueError(f"No schema definition found for JSON tool '{tool_name}'")
                 
-                # Use the clean_and_fix_json function from node_functions to handle potential JSON issues
-                from astonish.core.node_functions import clean_and_fix_json
+                # Use the clean_and_fix_json function to handle potential JSON issues
+                from astonish.core.json_utils import clean_and_fix_json
                 cleaned_json_string = clean_and_fix_json(input_string)
                 parsed_args = {} if not cleaned_json_string else json.loads(cleaned_json_string)
