+++ astonish/core/node_functions.py
@@ -1,3 +1,7 @@
+"""
+Node functions module for Astonish.
+This module contains functions for creating and managing node functions.
+"""
 import json
 import re
 import asyncio
@@ -5,7 +9,7 @@ import traceback
 import inquirer
 import readline
 import astonish.globals as globals
-from typing import TypedDict, Union, Optional, get_args, get_origin, Dict, Any, List, Callable, Coroutine, Type
+from typing import Dict, Any, Union, Optional, List
 from langchain_core.messages import SystemMessage, HumanMessage, BaseMessage
 from langchain_core.prompts import ChatPromptTemplate, BasePromptTemplate
 from langchain_core.tools import BaseTool
@@ -13,447 +17,18 @@ from langchain_core.runnables import Runnable
 from langchain_core.language_models.base import BaseLanguageModel
 from langchain.output_parsers import PydanticOutputParser
 from langchain.schema import OutputParserException
-from pydantic import Field, ValidationError, create_model, BaseModel
+from pydantic import ValidationError, BaseModel
 from astonish.tools.internal_tools import tools as internal_tools_list
 from astonish.core.llm_manager import LLMManager
 from astonish.core.utils import format_prompt, print_ai, print_output, console
 from astonish.core.error_handler import create_error_feedback, handle_node_failure
 from astonish.core.format_handler import execute_tool
 from astonish.core.utils import request_tool_execution
-from langchain_core.messages import SystemMessage, HumanMessage, BaseMessage
-from langchain_core.prompts import ChatPromptTemplate, BasePromptTemplate
-
-
-def clean_and_fix_json(content: str) -> str:
-    """
-    Clean and fix JSON content to make it parseable.
-    
-    Args:
-        content: The content to clean and fix
-        
-    Returns:
-        The cleaned and fixed content
-    """
-    if not content or not isinstance(content, str):
-        return ""
-
-    # 1. Attempt to find a JSON object or array using a regex.
-    # This is useful if the JSON is embedded within other text.
-    # Regex to find content starting with { or [ and ending with } or ] respectively,
-    # attempting to match balanced braces/brackets. This is a common heuristic.
-    # Note: A perfect regex for balanced structures is complex; this is an approximation.
-    json_match = re.search(r"(\{((?:[^{}]*|\{(?:[^{}]*|\{[^{}]*\})*\})*)\}|\[((?:[^\[\]]*|\[(?:[^\[\]]*|\[[^\[\]]*\])*\])*)\])", content)
-    
-    potential_json_str = ""
-    if json_match:
-        potential_json_str = json_match.group(0)
-        try:
-            # Try to parse to see if it's valid JSON
-            json.loads(potential_json_str)
-            # If it parses, we assume this is our primary JSON content
-            return potential_json_str.strip() 
-        except json.JSONDecodeError:
-            # If parsing fails, it might be that this regex match wasn't the actual JSON,
-            # or it's malformed. We'll proceed to other cleaning steps.
-            potential_json_str = "" # Reset, as this wasn't clean JSON
-
-    # 2. If no clear JSON structure was extracted via regex,
-    #    or if the extracted part was not valid, try cleaning common markdown wrappers
-    #    from the original content. This handles cases where the entire response
-    #    is meant to be JSON but is wrapped.
-    
-    cleaned_content = content.strip()
-    
-    # Remove ```json ... ``` or ``` ... ``` wrappers if they enclose the whole content
-    if cleaned_content.startswith("```json") and cleaned_content.endswith("```"):
-        cleaned_content = cleaned_content[len("```json") : -len("```")].strip()
-    elif cleaned_content.startswith("```") and cleaned_content.endswith("```"):
-        # If it's just ``` ... ```, remove them if the inner content looks like JSON
-        temp_inner = cleaned_content[len("```") : -len("```")].strip()
-        if temp_inner.startswith("{") or temp_inner.startswith("["):
-            cleaned_content = temp_inner
-        # else, it might be a non-JSON code block, so we probably shouldn't strip the ```
-        # and expect the parser to fail if it was expecting JSON.
-        # However, if potential_json_str was populated from regex and failed,
-        # cleaned_content is still the original content here. This part might need refinement
-        # based on how often LLMs return non-JSON in ``` when JSON is expected.
-        # For now, if regex found something, use that attempt; otherwise, use original cleaned_content.
-        
-    if potential_json_str and (not cleaned_content.startswith("{") and not cleaned_content.startswith("[")):
-        # This means regex found something, but it wasn't valid, and outer ``` stripping didn't apply
-        # or also resulted in non-JSON. This case is ambiguous.
-        # Let's prioritize the result of ``` stripping if it looks like JSON.
-        if cleaned_content.startswith("{") or cleaned_content.startswith("["):
-             return cleaned_content
-        # If regex found something that looked like JSON but wasn't, and ``` stripping didn't help,
-        # it's better to return the original stripped content and let the parser fail.
-        # However, for this specific function, we want the *most JSON-like* part.
-        # If no valid JSON was found, return the most "stripped" version assuming it was meant to be JSON.
-        return cleaned_content if (cleaned_content.startswith("{") or cleaned_content.startswith("[")) else potential_json_str
-
-    # If after all this, cleaned_content is still not starting with { or [,
-    # and potential_json_str (from regex) didn't validate, then it's unlikely to be the JSON we want.
-    # Return the most processed version that looks like JSON.
-    if cleaned_content.startswith("{") or cleaned_content.startswith("["):
-        return cleaned_content
-    elif potential_json_str: # from regex, even if it didn't validate alone
-        return potential_json_str
-
-    return content.strip() # Fallback to just stripped original content
-
-class ToolDefinition(TypedDict):
-    name: str
-    description: str
-    input_type: str
-    input_schema_definition: Optional[Any]
-    tool_executor: Callable[..., Coroutine[Any, Any, str]]
-    tool_instance: Optional[BaseTool]
-
-class ReactStepOutput(TypedDict):
-    """Structured output for a single ReAct planning step."""
-    status: str
-    tool: Optional[str]
-    tool_input: Optional[str]
-    answer: Optional[str]
-    thought: Optional[str]
-    raw_response: str
-    message_content_for_history: Optional[str]
-
-async def run_react_planning_step(
-    input_question: str,
-    agent_scratchpad: str,
-    system_message_content: str,
-    llm: BaseLanguageModel,
-    tool_definitions: List[ToolDefinition],
-    node_name: str = "ReAct Planner",
-    print_prompt: bool = False,
-) -> ReactStepOutput:
-    """
-    Performs one step of ReAct planning using a STRING scratchpad.
-    """
-    globals.logger.info(f"[{node_name}] Running ReAct planning step...")
-    default_error_output = ReactStepOutput(
-        status='error', tool=None, tool_input=None, answer=None, thought=None,
-        raw_response="Planner failed unexpectedly." # Removed message_content_for_history
-    )
-    try:
-        react_instructions_template = create_custom_react_prompt_template(tool_definitions)
-        prompt_str_template = system_message_content + "\n\n" + react_instructions_template
-        custom_react_prompt = ChatPromptTemplate.from_template(prompt_str_template)
-        chain: Runnable = custom_react_prompt | llm
-
-        invoke_input = {
-            "input": input_question,
-            "agent_scratchpad": agent_scratchpad
-        }
-
-        globals.logger.debug(f"[{node_name}] Invoking LLM with scratchpad:\n{agent_scratchpad}")
-        
-        response_chunks = []
-        async for chunk in chain.astream(invoke_input):
-            response_chunks.append(chunk)
-        
-        # Merge the chunks
-        if response_chunks:
-            # Create a complete response by concatenating all chunk contents
-            full_content = ""
-            for chunk in response_chunks:
-                if hasattr(chunk, 'content'):
-                    full_content += chunk.content
-            
-            # Use the last chunk as a template for the response object
-            llm_response = response_chunks[-1]
-            if hasattr(llm_response, 'content'):
-                llm_response.content = full_content
-        else:
-            llm_response = None
-            
-        response_text = llm_response.content if hasattr(llm_response, 'content') else str(llm_response)
-        globals.logger.info(f"[{node_name}] LLM Raw Planning Response:\n{response_text}")
-        if print_prompt:
-            print_output(invoke_input, "green")
-            #print_chat_prompt(ChatPromptTemplate(messages=messages), node_config)
-            #else: print_output(f"Cannot print prompt for {node_name}, 'messages' not defined.", "yellow")
-
-
-        thought_blocks = re.findall(r"^\s*Thought:\s*(.*?)(?=(?:\n\s*(?:Action:|Final Answer:|$)))", response_text, re.DOTALL | re.MULTILINE)
-        thought_text = thought_blocks[-1].strip() if thought_blocks else None
-
-        action_match = re.search(r"^\s*Action:\s*([\w.-]+)", response_text, re.MULTILINE | re.IGNORECASE)
-        input_string_from_llm = ""
-
-        action_input_line_match = re.search(r"^\s*Action Input:\s*(.*?)(?:\nObservation:|\Z)", response_text, re.DOTALL | re.MULTILINE | re.IGNORECASE)
-        if action_input_line_match:
-            raw_input_line = action_input_line_match.group(1).strip()
-
-            # Try to extract JSON content, or clean and fix the raw input
-            json_match = re.match(r"^```json\s*(\{.*?\})\s*```$", raw_input_line, re.DOTALL) or re.match(r"^(\{.*?\})\s*$", raw_input_line, re.DOTALL)
-            if json_match: 
-                input_string_from_llm = json_match.group(1)
-                globals.logger.debug(f"[{node_name}] Extracted JSON Action Input: {input_string_from_llm}")
-            else:
-                # If it looks like it might be JSON but didn't match the regex patterns, try to clean and fix it
-                if '{' in raw_input_line and '}' in raw_input_line:
-                    input_string_from_llm = clean_and_fix_json(raw_input_line)
-                    globals.logger.debug(f"[{node_name}] Cleaned and fixed JSON Action Input: {input_string_from_llm}")
-                else:
-                    input_string_from_llm = raw_input_line
-                    globals.logger.debug(f"[{node_name}] Extracted String Action Input: {input_string_from_llm}")
-
-        if action_match:
-            tool_name = action_match.group(1).strip()
-            globals.logger.info(f"[{node_name}] LLM planned Action: {tool_name}")
-
-            return ReactStepOutput(
-                status='action', tool=tool_name, tool_input=input_string_from_llm,
-                answer=None, thought=thought_text, raw_response=response_text,
-            )
-        elif "Final Answer:" in response_text:
-            final_answer_text = response_text.split("Final Answer:")[-1].strip()
-            globals.logger.info(f"[{node_name}] LLM provided Final Answer.")
-            return ReactStepOutput(
-                status='final_answer', tool=None, tool_input=None, answer=final_answer_text,
-                thought=thought_text, raw_response=response_text,
-            )
-        else:
-            error_message = f"LLM response did not contain 'Action:' or 'Final Answer:'."
-            globals.logger.error(f"[{node_name}] Parsing Error: {error_message}")
-            return ReactStepOutput(
-                status='error', tool=None, tool_input=None, answer=None, thought=thought_text,
-                raw_response=response_text,
-            )
-
-    except Exception as planning_error:
-        error_message = f"Critical error during ReAct planning step: {type(planning_error).__name__}: {planning_error}"
-        console.print(f"[{node_name}] {error_message}", style="red")
-        console.print(f"Traceback:\n{traceback.format_exc()}", style="red")
-        output = default_error_output.copy()
-        output['raw_response'] = f"Error: {error_message}\n{traceback.format_exc()}"
-        return output
-
-def create_output_model(output_model_config: Dict[str, str]) -> Optional[Type[BaseModel]]:
-    """
-    Creates a Pydantic model dynamically, handling lowercase type names from YAML
-    and complex/generic types using eval.
-    """
-    if not output_model_config or not isinstance(output_model_config, dict):
-        return None
-
-    fields = {}
-    type_lookup = {
-        "str": str,
-        "int": int,
-        "float": float,
-        "bool": bool,
-        "any": Any,
-        "list": List,
-        "dict": Dict,
-    }
-    eval_context = {
-         "Union": Union, "Optional": Optional, "List": List, "Dict": Dict,
-         "str": str, "int": int, "float": float, "bool": bool, "Any": Any,
-         "BaseModel": BaseModel
-         }
-
-    for field_name, field_type_str in output_model_config.items():
-        field_type = Any
-        try:
-            normalized_type_str = field_type_str.strip()
-            normalized_type_lower = normalized_type_str.lower()
-
-            if normalized_type_lower in type_lookup:
-                field_type = type_lookup[normalized_type_lower]
-
-            elif any(c in normalized_type_str for c in ['[', '|', 'Optional', 'Union']):
-                try:
-                     field_type = eval(normalized_type_str, globals(), eval_context)
-                except NameError:
-                     console.print(f"Warning: Eval failed to find type components within '{normalized_type_str}'. Defaulting field '{field_name}' to Any.", style="yellow")
-
-                     field_type = Any
-                except Exception as e_eval:
-                     console.print(f"Error evaluating complex type '{normalized_type_str}': {e_eval}", style="red")
-                     field_type = Any
-            else:
-                 console.print(f"Warning: Unknown or non-generic type '{normalized_type_str}' for field '{field_name}', defaulting to Any.", style="yellow")
-                 field_type = Any
-
-            if field_type is not Any:
-                origin = get_origin(field_type)
-                args = get_args(field_type)
-
-                if origin is Union and type(None) in args:
-                    non_none_args = tuple(arg for arg in args if arg is not type(None))
-                    if len(non_none_args) == 1: field_type = Optional[non_none_args[0]]
-                    elif len(non_none_args) > 1: field_type = Optional[Union[non_none_args]]
-                    else: field_type = Optional[type(None)]
-
-            fields[field_name] = (field_type, Field(description=f"{field_name} field"))
-
-        except Exception as e:
-            console.print(f"Error processing field '{field_name}' with type string '{field_type_str}': {e}", style="red")
-            fields[field_name] = (Any, Field(description=f"{field_name} field (processing error)"))
-
-    model_name = f"DynamicOutputModel_{abs(hash(json.dumps(output_model_config, sort_keys=True)))}"
-    try:
-         Model = create_model(model_name, **fields)
-         return Model
-    except Exception as e:
-         console.print(f"Failed to create Pydantic model '{model_name}': {e}", style="red")
-         return None
-
-def update_state(state: Dict[str, Any], output: Union[BaseModel, str, Dict, None], node_config: Dict[str, Any]) -> Dict[str, Any]:
-    """
-    Update the state with the output from a node execution.
-    
-    Args:
-        state: The current state dictionary
-        output: The output from the node execution
-        node_config: The node configuration dictionary
-        
-    Returns:
-        The updated state dictionary
-    """
-
-    if state.get('_error') is not None:
-        return state
-    
-    new_state = state.copy()
-    output_field_name = next(iter(node_config.get('output_model', {})), 'agent_final_answer')
-    
-    if isinstance(output, BaseModel):
-        new_state.update(output.model_dump(exclude_unset=True))
-    elif isinstance(output, dict):
-        if "_error" in output:
-            new_state['_error'] = output["_error"]
-            
-            if output["_error"] and not output["_error"].get('recoverable', True):
-                return {
-                    '_error': output["_error"],
-                    '_end': False
-                }
-
-        if "output" in output:
-            new_state[output_field_name] = output.get("output", "")
-    elif isinstance(output, str):
-        new_state[output_field_name] = output
-    elif output is None:
-        globals.logger.warning("Received None output for state update.")
-    
-    limit_counter_field = node_config.get('limit_counter_field')
-    limit = node_config.get('limit')
-    if limit_counter_field and limit and '_error' not in new_state:
-        counter = new_state.get(limit_counter_field, 0) + 1
-        if counter > limit:
-            counter = 1
-        new_state[limit_counter_field] = counter
-    
-    return new_state
-
-def create_custom_react_prompt_template(tools_definitions: List[ToolDefinition]) -> str:
-    """
-    Generates a ReAct prompt string including detailed tool input requirements,
-    WITH property descriptions, enums, defaults, and escaping schema braces.
-    Includes the {agent_scratchpad} placeholder for text-based history.
-    """
-    tool_strings = []
-    for tool_def in tools_definitions:
-        input_desc = ""
-        input_type = tool_def.get('input_type', 'STRING')
-        schema_def = tool_def.get('input_schema_definition')
-
-        if input_type == 'JSON_SCHEMA':
-            schema_desc_str = "[Schema not provided or invalid]"
-            properties = {}
-            required_list = []
-            if schema_def:
-                try:
-                    schema_json_props = None
-                    if isinstance(schema_def, type) and issubclass(schema_def, BaseModel):
-                        schema_json_props = schema_def.model_json_schema()
-                    elif isinstance(schema_def, dict) and "properties" in schema_def:
-                        schema_json_props = schema_def
-
-                    if schema_json_props:
-                        properties = schema_json_props.get("properties", {})
-                        required_list = schema_json_props.get("required", [])
-
-                    if properties:
-                        prop_details_list = []
-                        for name, details in properties.items():
-                            prop_type = details.get('type', 'unknown')
-                            prop_desc = details.get('description', '')
-                            prop_enum = details.get('enum')
-                            prop_default = details.get('default')
-
-                            is_required = name in required_list
-                            req_marker = " (required)" if is_required else ""
-
-                            detail_str = f"'{name}' ({prop_type}{req_marker})"
-                            if prop_desc:
-                                detail_str += f": {prop_desc}"
-                            if prop_enum:
-                                # Ensure enum values are correctly quoted if strings
-                                enum_strs = [f'"{v}"' if isinstance(v, str) else str(v) for v in prop_enum]
-                                detail_str += f" (must be one of: [{', '.join(enum_strs)}])" # List allowed values
-                            if prop_default is not None:
-                                 detail_str += f" (default: {json.dumps(prop_default)})" # Show default
-
-                            # Escape curly braces within the description details for f-string formatting
-                            prop_details_list.append(detail_str.replace("{", "{{").replace("}", "}}"))
-
-                        schema_desc_str = "; ".join(prop_details_list)
-                    else:
-                         schema_desc_str = "[No properties found in schema]"
-                except Exception as e:
-                    schema_desc_str = f"[Error extracting schema details]"
-
-            input_desc = f"Input Type: Requires a JSON object string with properties: {schema_desc_str}. IMPORTANT: Generate a valid JSON string containing required properties and only essential optional properties based on the user request."
-
-        elif input_type == 'STRING':
-             input_desc = "Input Type: Plain String"
-
-             if schema_def and isinstance(schema_def, str):
-                  # Escape curly braces in the example string
-                  escaped_example = schema_def.replace("{", "{{").replace("}", "}}")
-                  input_desc += f". Expected format/example: {escaped_example}"
-        else:
-             input_desc = f"Input Type: {input_type}"
-             if schema_def and isinstance(schema_def, str):
-                 # Escape curly braces
-                 input_desc += f". Tool expects: {schema_def.replace('{','{{').replace('}','}}')}"
-
-        tool_name = tool_def.get('name', 'UnnamedTool')
-        tool_description = tool_def.get('description', 'No description.')
-        # Escape braces in tool name and description just in case, though less likely needed
-        safe_tool_name = tool_name.replace("{", "{{").replace("}", "}}")
-        safe_tool_description = tool_description.replace("{", "{{").replace("}", "}}")
-        tool_strings.append(f"- {safe_tool_name}: {safe_tool_description} {input_desc}")
-
-    formatted_tools = "\n".join(tool_strings) if tool_strings else "No tools available."
-    tool_names = ", ".join([tool_def['name'] for tool_def in tools_definitions]) if tools_definitions else "N/A"
-    safe_tool_names = tool_names.replace("{", "{{").replace("}", "}}") # Escape braces in tool names list
-
-    template = f"""Answer the following questions as best you can. You have access to the following tools:
-        {formatted_tools}
-
-        Use the following format STRICTLY:
-
-        Question: the input question you must answer
-        Thought: Analyze the question and available tools. Determine the single best Action to take. Identify the essential arguments required by that Action's schema based on the Question and the tool description (especially paying attention to property descriptions, required fields, and allowed values like enums). Use sensible defaults for optional arguments unless the Question specifies otherwise.
-        Action: the action to take, must be one of [{safe_tool_names}]
-        Action Input: Provide the exact input required for the selected Action. If Input Type requires a JSON object string, generate a *single, valid JSON object string* containing ONLY the essential properties identified in your Thought process, matching the required types and allowed values (enums) mentioned in the tool description. If Input Type is STRING, provide the plain string. Do NOT add explanations before or after the Action Input line. IMPORTANT: After writing the Action Input line, STOP generating immediately. Wait for the Observation.
-        Observation: the result of the action
-        ... (this Thought/Action/Action Input/Observation can repeat N times)
-        Thought: I now know the final answer based on my thoughts and observations.
-        Final Answer: the final answer to the original input question. It must be the extraction of the content of the last Observation, without format modification, unless requested in the prompt.
-
-        Begin!
-
-        Question: {{input}}
-        {{agent_scratchpad}}"""
-
-    return template
+from astonish.core.json_utils import clean_and_fix_json
+from astonish.core.react_planning import ToolDefinition, ReactStepOutput, run_react_planning_step, format_react_step_for_scratchpad
+from astonish.core.output_model_utils import create_output_model, _format_final_output_with_llm
+from astonish.core.state_management import update_state
+from astonish.core.ui_utils import print_user_messages, print_chat_prompt, print_state
 
 def create_node_function(node_config, mcp_client):
     """Creates the appropriate node function based on node_config type."""
@@ -853,150 +428,3 @@ def create_llm_node_function(node_config: Dict[str, Any], mcp_client: Any, use_t
         return new_state
 
     return node_function
-
-def format_react_step_for_scratchpad(
-    thought: Optional[str],
-    action: Optional[str],
-    action_input: Optional[str],
-    observation: str # Observation is always required after an action
-    ) -> str:
-    """Formats a completed Thought-Action-Input-Observation step for the scratchpad string."""
-    # Ensure consistent spacing and newlines, similar to standard ReAct examples
-    scratchpad_entry = "\n" # Start with a newline for separation
-    if thought:
-        # Ensure thought doesn't already start with "Thought:" from LLM output
-        thought = re.sub(r"^\s*Thought:\s*", "", thought).strip()
-        scratchpad_entry += f"Thought: {thought}\n"
-    if action:
-         scratchpad_entry += f"Action: {action}\n"
-    if action_input is not None: # Can be empty string for actions without input
-         scratchpad_entry += f"Action Input: {action_input}\n"
-    # Observation comes from the tool execution result or denial
-    scratchpad_entry += f"Observation: {observation}\n"
-    return scratchpad_entry
-
-async def _format_final_output_with_llm(
-    final_text: str,
-    parser: PydanticOutputParser,
-    llm: BaseLanguageModel,
-    node_name: str
-) -> Union[BaseModel, str, Dict]: # Allow Dict if parser outputs dict
-    globals.logger.info(f"Attempting to format final result into JSON schema for {node_name}...")
-    format_instructions = ""
-    schema_valid_for_format = False
-    try:
-        if hasattr(parser.pydantic_object, 'model_json_schema'): 
-            format_instructions = parser.get_format_instructions()
-            schema_valid_for_format = True
-        else: 
-            console.print(f"Error: Pydantic V2 model lacks .model_json_schema() for {node_name}", style="red")
-    except Exception as e: 
-        console.print(f"[{node_name}] Unexpected error getting format instructions: {e}", style="red")
-    
-    if not schema_valid_for_format: 
-        console.print(f"Warning: Schema invalid or unavailable for formatting. Using raw ReAct result for {node_name}.", style="yellow")
-        return final_text
-    formatting_prompt = ( f"Analyze the following text, which is the final answer from a reasoning process. Your goal is to extract the single, most salient value or piece of information requested by the desired output schema (e.g., if the schema asks for 'execution_result', extract the primary outcome like a number, status, or summary; if it asks for 'weather_summary', extract that). Then, format ONLY this extracted value into a JSON object conforming strictly to the schema provided below. Respond ONLY with the JSON object, without any markdown formatting like ```json or explanations.\n\nFinal Answer Text:\n```\n{final_text}\n```\n\nRequired JSON Schema (extract the relevant value to fit this):\n{format_instructions}" )
-    formatting_messages = [HumanMessage(content=formatting_prompt)]
-    max_format_retries = 2
-    format_retry_count = 0
-    parsed_formatted_output = None
-    while format_retry_count < max_format_retries:
-        try:
-            globals.logger.info(f"Attempt {format_retry_count + 1}/{max_format_retries} for JSON formatting...")
-            response_chunks = []
-            async for chunk in llm.astream(formatting_messages):
-                response_chunks.append(chunk)
-            
-            # Merge the chunks
-            if response_chunks:
-                # Create a complete response by concatenating all chunk contents
-                full_content = ""
-                for chunk in response_chunks:
-                    if hasattr(chunk, 'content'):
-                        full_content += chunk.content
-                
-                # Use the last chunk as a template for the response object
-                formatting_response = response_chunks[-1]
-                if hasattr(formatting_response, 'content'):
-                    formatting_response.content = full_content
-            else:
-                formatting_response = None
-                
-            cleaned_content = clean_and_fix_json(formatting_response.content)
-            if not cleaned_content: raise OutputParserException("Received empty formatted response.")
-            parsed_formatted_output = parser.parse(cleaned_content)
-            globals.logger.info("Successfully extracted and formatted final result to JSON.")
-            return parsed_formatted_output
-        except (OutputParserException, ValidationError, json.JSONDecodeError) as parse_error:
-            format_retry_count += 1
-            error_detail = f"{type(parse_error).__name__}: {parse_error}"
-            console.print(f"Warning: Formatting LLM call failed parsing/validation (Attempt {format_retry_count}/{max_format_retries}): {error_detail}", style="yellow")
-            if format_retry_count >= max_format_retries: 
-                console.print(f"Failed to format final result to JSON after retries. Storing raw text result.", style="red")
-                return final_text
-            feedback = f"Formatting failed: {error_detail}. Extract the single key result value from the text and provide ONLY the valid JSON object matching the schema."
-            formatting_messages = [formatting_messages[0], HumanMessage(content=feedback)]
-        except Exception as format_error: 
-            console.print(f"Unexpected error during final result formatting LLM call: {format_error}", style="red")
-            return final_text
-    return final_text
-
-def print_user_messages(state: Dict[str, Any], node_config: Dict[str, Any]):
-    """Prints messages defined in node_config, substituting state variables."""
-    user_message_fields = node_config.get('user_message', [])
-    if not isinstance(user_message_fields, list):
-         print(f"Warning: 'user_message' in node config is not a list.", style="yellow")
-         return
-    for field_or_template in user_message_fields:
-        if isinstance(field_or_template, str):
-            if field_or_template in state and state[field_or_template] is not None:
-                 print_ai(str(state[field_or_template])) # Print the value directly
-                 continue
-
-            try:
-                 formatted_msg = format_prompt(field_or_template, state, node_config)
-                 print_ai(formatted_msg)
-            except Exception as e:
-                 console.print(f"Error formatting user message '{field_or_template}': {e}", style="red")
-                 print_ai(field_or_template) # Print original on error
-        else:
-             console.print(f"Warning: Item in 'user_message' is not a string: {field_or_template}", style="yellow")
-
-def print_chat_prompt(chat_prompt: BasePromptTemplate, node_config: Dict[str, Any]):
-    """Prints formatted chat prompt messages if enabled in config."""
-    print_prompt_flag = node_config.get('print_prompt', False)
-    node_name = node_config.get('name', 'Unknown Node')
-    if print_prompt_flag and hasattr(chat_prompt, 'messages'):
-        console.print(f"ChatPrompt for {node_name}:", style="blue")
-        try:
-            for i, message in enumerate(chat_prompt.messages, 1):
-                role = "Unknown"
-                content = str(message)
-                color = "red"
-                
-                if isinstance(message, SystemMessage): 
-                    role = "System"
-                    content = message.content
-                    color = "magenta"
-                elif isinstance(message, HumanMessage): 
-                    role = "Human"
-                    content = message.content
-                    color = "yellow"
-                content_preview = content
-                print(f"  {color}{role} {i}: {content_preview}")
-        except Exception as e: console.print(f"Error printing chat prompt: {e}", style="red")
-        print("-" * 20)
-
-
-def print_state(state: Dict[str, Any], node_config: Dict[str, Any]):
-    """Prints the current state dictionary if enabled in config."""
-    print_state_flag = node_config.get('print_state', False)
-    node_name = node_config.get('name', 'Unknown Node')
-    if print_state_flag:
-        print_output(f"Current State after {node_name}:")
-        try:
-            state_str = json.dumps(state, indent=2, default=lambda o: f"<non-serializable: {type(o).__name__}>")
-            console.print(f"{state_str}", style="green")
-        except Exception as e: print(f"Could not serialize state to JSON. Error: {e}. Raw state:\n{state}", style="red")
-        print("-" * 20)
