+++ astonish/core/prompt_templates.py
@@ -1,0 +1,112 @@
+"""
+Prompt template utilities for Astonish.
+This module contains functions for creating and managing prompt templates.
+"""
+import json
+from typing import List, Dict, Any
+from pydantic import BaseModel
+
+def create_custom_react_prompt_template(tools_definitions: List[Dict[str, Any]]) -> str:
+    """
+    Generates a ReAct prompt string including detailed tool input requirements,
+    WITH property descriptions, enums, defaults, and escaping schema braces.
+    Includes the {agent_scratchpad} placeholder for text-based history.
+    """
+    tool_strings = []
+    for tool_def in tools_definitions:
+        input_desc = ""
+        input_type = tool_def.get('input_type', 'STRING')
+        schema_def = tool_def.get('input_schema_definition')
+
+        if input_type == 'JSON_SCHEMA':
+            schema_desc_str = "[Schema not provided or invalid]"
+            properties = {}
+            required_list = []
+            if schema_def:
+                try:
+                    schema_json_props = None
+                    if isinstance(schema_def, type) and issubclass(schema_def, BaseModel):
+                        schema_json_props = schema_def.model_json_schema()
+                    elif isinstance(schema_def, dict) and "properties" in schema_def:
+                        schema_json_props = schema_def
+
+                    if schema_json_props:
+                        properties = schema_json_props.get("properties", {})
+                        required_list = schema_json_props.get("required", [])
+
+                    if properties:
+                        prop_details_list = []
+                        for name, details in properties.items():
+                            prop_type = details.get('type', 'unknown')
+                            prop_desc = details.get('description', '')
+                            prop_enum = details.get('enum')
+                            prop_default = details.get('default')
+
+                            is_required = name in required_list
+                            req_marker = " (required)" if is_required else ""
+
+                            detail_str = f"'{name}' ({prop_type}{req_marker})"
+                            if prop_desc:
+                                detail_str += f": {prop_desc}"
+                            if prop_enum:
+                                # Ensure enum values are correctly quoted if strings
+                                enum_strs = [f'"{v}"' if isinstance(v, str) else str(v) for v in prop_enum]
+                                detail_str += f" (must be one of: [{', '.join(enum_strs)}])" # List allowed values
+                            if prop_default is not None:
+                                 detail_str += f" (default: {json.dumps(prop_default)})" # Show default
+
+                            # Escape curly braces within the description details for f-string formatting
+                            prop_details_list.append(detail_str.replace("{", "{{").replace("}", "}}"))
+
+                        schema_desc_str = "; ".join(prop_details_list)
+                    else:
+                         schema_desc_str = "[No properties found in schema]"
+                except Exception as e:
+                    schema_desc_str = f"[Error extracting schema details]"
+
+            input_desc = f"Input Type: Requires a JSON object string with properties: {schema_desc_str}. IMPORTANT: Generate a valid JSON string containing required properties and only essential optional properties based on the user request."
+
+        elif input_type == 'STRING':
+             input_desc = "Input Type: Plain String"
+
+             if schema_def and isinstance(schema_def, str):
+                  # Escape curly braces in the example string
+                  escaped_example = schema_def.replace("{", "{{").replace("}", "}}")
+                  input_desc += f". Expected format/example: {escaped_example}"
+        else:
+             input_desc = f"Input Type: {input_type}"
+             if schema_def and isinstance(schema_def, str):
+                 # Escape curly braces
+                 input_desc += f". Tool expects: {schema_def.replace('{','{{').replace('}','}}')}"
+
+        tool_name = tool_def.get('name', 'UnnamedTool')
+        tool_description = tool_def.get('description', 'No description.')
+        # Escape braces in tool name and description just in case, though less likely needed
+        safe_tool_name = tool_name.replace("{", "{{").replace("}", "}}")
+        safe_tool_description = tool_description.replace("{", "{{").replace("}", "}}")
+        tool_strings.append(f"- {safe_tool_name}: {safe_tool_description} {input_desc}")
+
+    formatted_tools = "\n".join(tool_strings) if tool_strings else "No tools available."
+    tool_names = ", ".join([tool_def['name'] for tool_def in tools_definitions]) if tools_definitions else "N/A"
+    safe_tool_names = tool_names.replace("{", "{{").replace("}", "}}") # Escape braces in tool names list
+
+    template = f"""Answer the following questions as best you can. You have access to the following tools:
+        {formatted_tools}
+
+        Use the following format STRICTLY:
+
+        Question: the input question you must answer
+        Thought: Analyze the question and available tools. Determine the single best Action to take. Identify the essential arguments required by that Action's schema based on the Question and the tool description (especially paying attention to property descriptions, required fields, and allowed values like enums). Use sensible defaults for optional arguments unless the Question specifies otherwise.
+        Action: the action to take, must be one of [{safe_tool_names}]
+        Action Input: Provide the exact input required for the selected Action. If Input Type requires a JSON object string, generate a *single, valid JSON object string* containing ONLY the essential properties identified in your Thought process, matching the required types and allowed values (enums) mentioned in the tool description. If Input Type is STRING, provide the plain string. Do NOT add explanations before or after the Action Input line. IMPORTANT: After writing the Action Input line, STOP generating immediately. Wait for the Observation.
+        Observation: the result of the action
+        ... (this Thought/Action/Action Input/Observation can repeat N times)
+        Thought: I now know the final answer based on my thoughts and observations.
+        Final Answer: the final answer to the original input question. It must be the extraction of the content of the last Observation, without format modification, unless requested in the prompt.
+
+        Begin!
+
+        Question: {{input}}
+        {{agent_scratchpad}}"""
+
+    return template
